travel_agent/requirements.txt
google-adk==1.6.1
toolbox-langchain
langchain
a2a-sdk
